Python/serial_eval/requirements.txt
astroid==2.3.3
crc16==0.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.4.4
pyserial==3.4
six==1.14.0
wrapt==1.11.2
